Add files via upload
these files are from almost a year ago since then I have learned some lessons like to document what something is and why it is there if there is something that is accessed by multiple things write it down somewhere what uses it. along with the do it right the first time don't just finish something quick to feel-good work on it till it is done right.